Manage table create and destroy
